finish the gradient algorithms
